+++ spec/scss_lint/rake_task_spec.rb
@@ -1,20 +1,46 @@
 require 'spec_helper'
 require 'scss_lint/rake_task'
+require 'tempfile'
 
 describe SCSSLint::RakeTask do
-  before do
-    # Silence console output
-    STDOUT.stub(:write)
+  before(:all) do
+    SCSSLint::RakeTask.new
   end
 
-  describe '#run' do
-    subject do
-      SCSSLint::RakeTask.new
-      Rake::Task['scss_lint']
-    end
+  before do
+    STDOUT.stub(:write) # Silence console output
+  end
 
-    it 'raises error when no files are specified' do
-      lambda { subject.invoke.first.call }.should raise_error SystemExit
+  let(:file) do
+    Tempfile.new(%w[scss-file .scss]).tap do |f|
+      f.write(scss)
+      f.close
+    end
+  end
+
+  def run_task_and_get_status
+    Rake::Task[:scss_lint].reenable # Allows us to execute task multiple times
+
+    begin
+      Rake::Task[:scss_lint].invoke(file.path)
+    rescue SystemExit => ex
+      ex.status
+    end
+  end
+
+  context 'when SCSS document is valid with no lints' do
+    let(:scss) { '' }
+
+    it 'executes successfully' do
+      run_task_and_get_status.should == 0
+    end
+  end
+
+  context 'when SCSS document is invalid' do
+    let(:scss) { '.class {' }
+
+    it 'returns an error status code' do
+      run_task_and_get_status.should == 2
     end
   end
 end